+++ hamplicons/__init__.py
@@ -1,2 +1,0 @@
-__version_tuple__ = (4, 1, 1)
-__version__ = ".".join(map(str, __version_tuple__))