+++ tests/test_normal.py
@@ -1,0 +1,9 @@
+from hamplicons.main import main, parse_args
+
+
+class TestClass:
+	def test_total(self):
+		main(parse_args())
+
+
+