+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changelog
 
+
+## [4.2.0] - 2022-09-22
+  - Added pytest tests
+  - Added option '...' to output indel size data
+  - Updated dependencies
+
 ## [4.1.0] - 2020-11-07
 
 ### Added